backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymupdf==1.23.7
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
requests==2.31.0
aiofiles==23.2.1
python-magic==0.4.27
pydantic==2.5.0